% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{logrittr_activate}
\alias{logrittr_activate}
\title{Activate logrittr logging on the \%>\% pipe}
\usage{
logrittr_activate()
}
\value{
Invisibly returns the previous definition of \verb{\%>\%} in the global
environment (or \code{NULL} if none existed).
}
\description{
Replaces \verb{\%>\%} in the global environment with \verb{\%>=\%} so that existing
pipelines using \verb{\%>\%} are automatically logged without any code change.

Call \code{\link[=logrittr_deactivate]{logrittr_deactivate()}} to restore the original \verb{\%>\%} behaviour.
}
\examples{
\dontrun{
library(logrittr)
library(dplyr)

logrittr_activate()

iris \%>\%
  filter(Sepal.Length < 5) \%>\%
  group_by(Species) \%>\%
  summarise(n = n())

logrittr_deactivate()
}

}
\seealso{
\code{\link[=logrittr_deactivate]{logrittr_deactivate()}}, \code{\link[=logrittr_hook]{logrittr_hook()}}
}
